refactor: change directory some files
