+++ src/components/ImageCardContainer/ImageCard/ImageCard.scss
