+++ src/components/ImageCardContainer/ImageCard/ImageCardPopup/ImageCardPopup.scss
