+++ src/components/ImageCardContainer/ImageCard/ImageCardPopup/ImageCardPopup.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, render, screen } from '@testing-library/react';
 import ImageCard from '../ImageCard';
-import { catQueryMock } from '../../../mocks/catQueryMock';
+import { catQueryMock } from '../../../../mocks/catQueryMock';
 import React from 'react';
 
 it('Test image card popup', async () => {
